pyproject.toml
[project]
name = "unsloth-grpo"
version = "0.1.0"
description = "GRPO training for Qwen3-4B-Thinking with unsloth"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "unsloth",
    "vllm>=0.8.0",
    "trl>=0.22.0",
    "datasets",
    "torch",
    "transformers>=4.56.0",
    "accelerate",
    "bitsandbytes",
    "wandb",
    "openai",
    "python-dotenv",
    "scikit-learn",
    "pandas",
    "inspect-ai",
    "anthropic>=0.75.0",
    "matplotlib>=3.10.8",
]

[tool.setuptools.packages.find]
where = ["src"]
